Update documentation for Chrome M118
After M118, Chrome now supports Passkeys on macOS from iCloud Keychain,
as long as macOS 13.5 or later is installed. I've personally tested all
of the relevant flows and seen them working. I would appreciate another
set of eyes to verify this, of course.

The underlying source on the web for this is
https://bugs.chromium.org/p/chromium/issues/detail?id=1439987.